package 'client.model.topic' moved to 'client.model.local.topic'
