+++ src/main/java/egl/client/controller/task/category/MissingLettersTaskController.java
@@ -8,8 +8,8 @@ import java.util.Random;
 import java.util.ResourceBundle;
 
 import egl.client.controller.task.AbstractTaskController;
-import egl.client.model.topic.category.Category;
-import egl.client.model.topic.category.Translation;
+import egl.client.model.local.topic.category.Category;
+import egl.client.model.local.topic.category.Translation;
 import egl.client.view.pane.CustomBorderPane;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
