+++ src/main/java/egl/client/controller/topic/CategoryInfoController.java
@@ -3,8 +3,8 @@ package egl.client.controller.topic;
 import egl.client.controller.info.AbstractEntityInfoController;
 import egl.client.controller.topic.info.TheoryTextArea;
 import egl.client.controller.topic.info.TranslationsEditableListView;
-import egl.client.model.topic.category.Category;
-import egl.client.model.topic.category.Translation;
+import egl.client.model.local.topic.category.Category;
+import egl.client.model.local.topic.category.Translation;
 import egl.client.view.info.NameDescriptionInfoView;
 import egl.client.view.pane.CustomBorderPane;
 import javafx.fxml.FXML;
