+++ src/main/java/egl/client/controller/topic/LocalTopicsController.java
@@ -4,8 +4,8 @@ import egl.client.controller.Controller;
 import egl.client.controller.WindowController;
 import egl.client.controller.profile.GlobalProfileController;
 import egl.client.controller.profile.LocalProfilesController;
-import egl.client.model.topic.LocalTopic;
-import egl.client.model.topic.category.Category;
+import egl.client.model.local.topic.LocalTopic;
+import egl.client.model.local.topic.category.Category;
 import egl.client.service.FxmlService;
 import egl.client.service.model.profile.GlobalProfileService;
 import egl.client.service.model.profile.LocalProfileService;
