+++ src/main/java/egl/client/controller/topic/info/TranslationsListView.java
@@ -1,7 +1,7 @@
 package egl.client.controller.topic.info;
 
-import egl.client.model.topic.category.Category;
-import egl.client.model.topic.category.Translation;
+import egl.client.model.local.topic.category.Category;
+import egl.client.model.local.topic.category.Translation;
 import egl.client.service.FxmlService;
 import egl.client.view.info.EntityInfoView;
 import egl.client.view.table.list.CustomListView;
