+++ src/main/java/egl/client/model/local/topic/category/Language.java
@@ -1,4 +1,4 @@
-package egl.client.model.topic.category;
+package egl.client.model.local.topic.category;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
