+++ src/main/java/egl/client/repository/topic/CategoryRepository.java
@@ -1,6 +1,6 @@
 package egl.client.repository.topic;
 
-import egl.client.model.topic.category.Category;
+import egl.client.model.local.topic.category.Category;
 import egl.client.repository.EntityRepository;
 import org.springframework.stereotype.Repository;
