+++ src/main/java/egl/client/service/model/topic/CategoryService.java
@@ -1,6 +1,6 @@
 package egl.client.service.model.topic;
 
-import egl.client.model.topic.category.Category;
+import egl.client.model.local.topic.category.Category;
 import egl.client.repository.topic.CategoryRepository;
 import egl.client.service.model.AbstractEntityService;
 import egl.core.model.topic.TopicType;
